Refactor import statements for Conv2dBlock in test_conv.py and test_resnet18.py
