+++ deeplay/tests/models/backbones/test_resnet18.py
@@ -3,7 +3,7 @@ import torch
 import torch.nn as nn
 
 
-from deeplay.blocks.conv2d.conv import Conv2dBlock
+from deeplay.blocks.conv.conv2d import Conv2dBlock
 from deeplay.models.backbones.resnet18 import BackboneResnet18
